allow backup of our own package, but do not stop it
